ref: move MAX_RELAYS_ERROR to utils/constants
