+++ src/services/pocket-relayer.ts
@@ -7,7 +7,8 @@ import { HttpErrors } from '@loopback/rest'
 import { MetricsRecorder } from '../services/metrics-recorder'
 import { PocketAAT, Session, RelayResponse, Pocket, Configuration, HTTPMethod, Node } from '@pokt-network/pocket-js'
 import { Redis } from 'ioredis'
-import { RelayError, MAX_RELAYS_ERROR } from '../errors/types'
+import { RelayError } from '../errors/types'
+import { MAX_RELAYS_ERROR } from '../utils/constants'
 import AatPlans from '../config/aat-plans.json'
 
 import { JSONObject } from '@loopback/context'
