+++ src/utils/constants.ts
@@ -5,3 +5,5 @@ export const WS_ONLY_METHODS = [
   'eth_getFilterChanges',
   'eth_getFilterLogs',
 ]
+
+export const MAX_RELAYS_ERROR = 'the evidence is sealed, either max relays reached or claim already submitted'
