+++ tests/unit/pocket-relayer.unit.ts
@@ -17,7 +17,8 @@ import { metricsRecorderMock } from '../mocks/metricsRecorder'
 import { DEFAULT_NODES, PocketMock } from '../mocks/pocketjs'
 import { BlockchainsRepository } from '../../src/repositories/blockchains.repository'
 import { gatewayTestDB } from '../fixtures/test.datasource'
-import { LimitError, MAX_RELAYS_ERROR } from '../../src/errors/types'
+import { LimitError } from '../../src/errors/types'
+import { MAX_RELAYS_ERROR } from '../../src/utils/constants'
 import { checkWhitelist, parseMethod } from '../../src/utils/string'
 import { checkSecretKey, updateConfiguration } from '../../src/utils/pocket'
 import { loadBlockchain } from '../../src/utils/relayer'
